Merge the following changes (catch-funcalls).
2009-08-19  Raphael 'kena' Poss  <r.c.poss@uva.nl>

* tools/lib/mtalpha-sim/callgate.s: Simplify.

2009-08-19  Raphael 'kena' Poss  <r.c.poss@uva.nl>

More support for function calls.

* tools/bin/spp: Try hard to avoid spilling "preserved" registers
to the frame. Try harder to remove frame setup and restore upon
end of thread.

* tools/lib/t-pppalpha/slimpl.m4: Pass function calls directly.

2009-08-19  Raphael 'kena' Poss  <r.c.poss@uva.nl>

Reduce support to function calls only.

The previous changes would also rewrite calls to the divide
routines; but this is not needed any more since the decision was
made separately to implement a hardware integer divide instead.

* tools/bin/spp: Reinstate the use of t10 and t11 for parameter
passing.

* tools/lib/t-pppalpha/include/sl_support.h: Lower the maximum
number of locals again.

2009-08-18  Raphael 'kena' Poss  <r.c.poss@uva.nl>

Preliminary support for rewriting function calls in assembly.

* tools/bin/spp,
* tools/lib/mtalpha-sim/callgate.s,
* tools/lib/t-pppalpha/include/sl_support.h: New features.