[sl-core] Make package more reusable for other trees.
There are two ways the parts of the source tree can be used:
either together as sub-trees of the "base" package, or
independently of the base package after the `sl-core' package was
installed somewhere on the system.

In the former case the tools and makefile rules can be picked up
relative to the directory where they are reused; in the latter
case they must pick them up from some agreed-upon install
location.

The new `sl-init-autoconf' script implements this protocol by
fetching the aclocal and Makefile rules (.mk) from the "nearest"
location, either a source tree "nearby" (next to the
`sl-init-autoconf' script itself) or from the install locations.

* build-aux/sl-init-autoconf: New.

* build-aux/Makefile.am: Install `sl-init-autoconf'. Also install
the `.mk' files to some location where `sl-init-autoconf' can find
them.

* Makefile.am: Recurse into `build-aux'.