[sl-core] Fix bug in MT-Alpha code generation.
The thread index would not be set properly if the `sl_index'
declaration was placed later than the start of the thread function
body (e.g. after a function call). Reported by M. Hicks and
M. Lankamp.

* tools/lib/slc/mtalpha/visitors.py (TFun_2_MTATFun.visit_fundef),
* tools/lib/t-sppalpha/include/sl_support.h (__sl_declindex):
Force a dependency on the `asm' construct to avoid (erroneous)
elision by the code generator.